+++ CHANGELOG.md
@@ -4,6 +4,10 @@
 ### 🐞 Bug Fixes
 * ColumnSpec.actions accepts Array<RecordActionSpec|RecordAction>
 
+### 📚 Libraries
+
+* onsenui `2.11.1 -> 2.12.8`
+* react-onsenui `1.11.5 > 1.13.2`
 
 ### ⚙️ Technical